TagGroupChanges and Notifications JSON
Note that I need to check over the last one a bit.
I'm going to be refactoring notifications quite heavily to be more
de-normalized in the near future, which should speed up fetches at the
price of taking up more DB space. So we'll have to see exactly what
happens there.