User sets content pref properly
This was caused by a lot of things. Basically, the check boxes get sent
as the strings "0" for false and "1" for true. This is OK when used for
boolean fields, since rails converts those automatically, but bad when
used for the user's actual content preference, as they get saved as the
raw strings. Strings in ruby are always true, and it's a pain in the ass
to manually check every time, so we just coerece the types back in an
after_save callback on user.

Hopefully that makes sense.